Pull up following revision(s) (requested by mlelstv in ticket #261): 	sys/dev/sdmmc/ld_sdmmc.c: revision 1.32 	sys/dev/sdmmc/ld_sdmmc.c: revision 1.33 	sys/dev/sdmmc/ld_sdmmc.c: revision 1.34 	sys/dev/sdmmc/sdmmc_mem.c: revision 1.62 	sys/dev/i2o/ld_iop.c: revision 1.39 	sys/dev/ld.c: revision 1.102 	sys/dev/ld.c: revision 1.103 	sys/dev/dksubr.c: revision 1.98 	sys/dev/dksubr.c: revision 1.99 	sys/dev/sdmmc/sdmmcvar.h: revision 1.29 	sys/dev/ic/ld_nvme.c: revision 1.17 	sys/dev/ldvar.h: revision 1.31 	sys/dev/ldvar.h: revision 1.32 	sys/dev/ic/ld_cac.c: revision 1.31 	sys/dev/pci/ld_virtio.c: revision 1.16 While ld(4) is MP safe, many backends are not. Add a flag for backends that are MP safe. Take KERNEL_LOCK when calling into a backend that doesn't have the flag set. Do the same for the discard routine. Fixes PR 52462. Defer sdmmc discard operations to the sdmmc task queue. Fixes a panic introduced by ld.c r1.102. validate length for discard operation and split operation when byte length doesn't fit into 'int'. make the sc_discard interface for the ld backend asynchronous and signal completion through new callback lddiscardend. Use a standard struct buf to pass disk address and range instead of two off_t values. make lddiscard synchronous again. This is a requirement of the current ffs discard code. Initialize error also in the case where len=0, which just succeeds. while here, assert that the len is indeed non-negative. this is already confirmed by sys_fdiscard, but let's be sure. reported by: GCC, but with different compile flags

+++ sys/dev/sdmmc/sdmmcvar.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: sdmmcvar.h,v 1.23.6.3 2017/07/25 01:49:13 snj Exp $	*/
+/*	$NetBSD: sdmmcvar.h,v 1.23.6.4 2017/09/01 09:59:10 martin Exp $	*/
 /*	$OpenBSD: sdmmcvar.h,v 1.13 2009/01/09 10:55:22 jsg Exp $	*/
 
 /*
@@ -381,7 +381,7 @@ int	sdmmc_mem_read_block(struct sdmmc_function *, uint32_t, u_char *,
 	    size_t);
 int	sdmmc_mem_write_block(struct sdmmc_function *, uint32_t, u_char *,
 	    size_t);
-int	sdmmc_mem_discard(struct sdmmc_function *, off_t, off_t);
+int	sdmmc_mem_discard(struct sdmmc_function *, uint32_t, uint32_t);
 int	sdmmc_mem_flush_cache(struct sdmmc_function *, bool);
 
 #endif	/* _SDMMCVAR_H_ */
